requirements.txt
# Data Handling
pandas
lxml
openpyxl

# API & Web
fastapi
uvicorn[standard]
streamlit

# Config & Validation
pydantic
pyyaml

# File uploads
python-multipart

# Testing
pytest

